scripts/requirements.txt
gtts
playsound
pyttsx3
boto3
paramiko
scp